requirements.txt
anyio>=3.6.2
asgiref>=3.6.0
certifi>=2022.12.7
cffi>=1.15.1
charset-normalizer>=3.1.0
click>=8.1.3
cryptography>=40.0.1
ecdsa>=0.18.0
fastapi>=0.95.0
gunicorn>=20.1.0
h11>=0.14.0
httpcore>=0.16.3
httpx>=0.23.3
idna>=3.4
pyasn1>=0.4.8
pycparser>=2.21
pydantic>=1.10.7
PyJWT>=2.6.0
python-jose>=3.3.0
rfc3986>=1.5.0
rsa>=4.9
six>=1.16.0
sniffio>=1.3.0
starlette>=0.26.1
typing_extensions>=4.5.0
uvicorn>=0.21.1
fastapi-sessions>=0.3.2
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.2
